requirements.txt
paramiko
ShopifyAPI
python-dateutil
xmlformatter
python_graphql_client
python-slugify
unicode

